1. Removing sortable-item-mixin. (#270)
2. Combined sortable-item tests with sortable-item-mixin tests.
3. Moved functions outside of component into utils
