+++ addon/utils/constant.js
@@ -1,0 +1,3 @@
+export const DRAG_ACTIONS = ['mousemove', 'touchmove'];
+export const ELEMENT_CLICK_ACTION = 'click';
+export const END_ACTIONS = ['click', 'mouseup', 'touchend'];
