+++ addon/utils/css-calculation.js
@@ -1,0 +1,17 @@
+/**
+  Gets a numeric border-spacing values for a given element.
+
+  @method getBorderSpacing
+  @param {Element} element
+  @return {Object}
+  @private
+*/
+export function getBorderSpacing(el) {
+  let css = getComputedStyle(el).borderSpacing; // '0px 0px'
+  let [horizontal, vertical] = css.split(' ');
+
+  return {
+    horizontal: parseFloat(horizontal),
+    vertical: parseFloat(vertical)
+  };
+}
