--- tests/dummy/app/components/sortable-item-mixin.js
@@ -1,4 +1,0 @@
-import Component from '@ember/component';
-import SortableItemMixin from 'ember-sortable/mixins/sortable-item';
-
-export default Component.extend(SortableItemMixin);
